1. For this requirement, the service already forces unique IDs so it is not checked. A new visitor was created to go through the IDs and check them for spaces. A button was added to the main control panel which will run this check. The result is then displayed in a dialog window.
2. For this requirement, I turned the GroupComponents into an Abstract
class. When a componenet is created, such as a user or a group, the
creation time is printed to the console, and it is also stored.

3. This requirement is implemented in the User class. It updates when the
user is first created, and anytime a user or someone he/she is following
post a tweet. In the User View, this time is shown and updated
accordingly.

4. A button was added to the main control panel. This button creates a
visitor which visits all user. The visitor stores a reference to the most
recently updated user. That user is compared to other users, and is
replaced if there is a newer update.